* master: (220 commits)
  update version
  modify  usms sendmessage test
  modify usms SendUSMSMessageParam.TemplateParams is not required
  update md
  modify ObjectToParam.objectToParams to support get param from sub class
  remove unused import and unused code
  update version
  UDPN.DescribeUDPNResult.UDPNData add field
  UNET.DescribeEIPResult增加属性
  UNET.DescribeShareBandwidthResult.ShareBandwidth增加属性
  UNET.GetEIPUpgradePriceResult增加属性
  UNET.GetEIPPriceResult.EIPPrice增加属性
  ULB和UNET部分参数更新
  修改参数对象转化成参数的逻辑，支持多层继承
  add ubill module
  update http request log format
  update http request log
  update http request log
  update usms file read
  update md
  ...

# Conflicts:
#	.gitignore
#	.gitlab-ci.yml
#	README.md
#	pom.xml
#	ucloud-sdk-java-censor/pom.xml
#	ucloud-sdk-java-censor/src/main/java/cn/ucloud/censor/client/DefaultCensorClient.java
#	ucloud-sdk-java-common/pom.xml
#	ucloud-sdk-java-common/src/main/java/cn/ucloud/common/client/UcloudClient.java
#	ucloud-sdk-java-common/src/main/java/cn/ucloud/common/http/Http.java
#	ucloud-sdk-java-common/src/main/java/cn/ucloud/common/http/UcloudHttpImpl.java
#	ucloud-sdk-java-ocr/pom.xml
#	ucloud-sdk-java-ocr/src/main/java/cn/ucloud/ocr/client/DefaultOcrClient.java
#	ucloud-sdk-java-pathx/pom.xml
#	ucloud-sdk-java-pathx/src/main/java/cn/ucloud/pathx/client/DefaultPATHXClient.java
#	ucloud-sdk-java-pathx/src/main/java/cn/ucloud/pathx/model/DescribeUGAInstanceResult.java
#	ucloud-sdk-java-pathx/src/main/java/cn/ucloud/pathx/model/GetCommonUGAPriceResult.java
#	ucloud-sdk-java-pathx/src/main/java/cn/ucloud/pathx/model/UGAAInfo.java
#	ucloud-sdk-java-pathx/src/main/java/cn/ucloud/pathx/model/UGAALine.java
#	ucloud-sdk-java-pathx/src/main/java/cn/ucloud/pathx/model/UGAATask.java
#	ucloud-sdk-java-pathx/src/main/java/cn/ucloud/pathx/model/UPathInfo.java
#	ucloud-sdk-java-pathx/src/main/java/cn/ucloud/pathx/model/UPathSet.java
#	ucloud-sdk-java-udb/pom.xml
#	ucloud-sdk-java-udb/src/main/java/cn/ucloud/udb/client/DefaultUDBClient.java
#	ucloud-sdk-java-udb/src/main/java/cn/ucloud/udb/model/DescribeUDBInstanceResult.java
#	ucloud-sdk-java-udisk/pom.xml
#	ucloud-sdk-java-udisk/src/main/java/cn/ucloud/udisk/model/DescribeUDiskResult.java
#	ucloud-sdk-java-udpn/pom.xml
#	ucloud-sdk-java-ufs/pom.xml
#	ucloud-sdk-java-ufs/src/test/java/cn/ucloud/ufs/client/CreateUFSVolumeTest.java
#	ucloud-sdk-java-ufs/src/test/java/cn/ucloud/ufs/client/DescribeUFSVolumePriceTest.java
#	ucloud-sdk-java-ufs/src/test/java/cn/ucloud/ufs/client/DescribeUFSVolumeUpgradePriceTest.java
#	ucloud-sdk-java-ufs/src/test/java/cn/ucloud/ufs/client/ExtendUFSVolumeTest.java
#	ucloud-sdk-java-uhost/pom.xml
#	ucloud-sdk-java-uhost/src/main/java/cn/ucloud/uhost/client/DefaultUhostClient.java
#	ucloud-sdk-java-uhost/src/main/java/cn/ucloud/uhost/model/DescribeUHostInstanceResult.java
#	ucloud-sdk-java-uhost/src/main/java/cn/ucloud/uhost/model/GetUHostInstancePriceResult.java
#	ucloud-sdk-java-ulb/pom.xml
#	ucloud-sdk-java-ulb/src/main/java/cn/ucloud/ulb/model/DescribeVServerResult.java
#	ucloud-sdk-java-umem/pom.xml
#	ucloud-sdk-java-umem/src/main/java/cn/ucloud/umem/client/DefaultUMEMClient.java
#	ucloud-sdk-java-umem/src/main/java/cn/ucloud/umem/model/DescribeUMemcacheUpgradePriceResult.java
#	ucloud-sdk-java-umem/src/main/java/cn/ucloud/umem/model/DescribeURedisUpgradePriceResult.java
#	ucloud-sdk-java-umem/src/main/java/cn/ucloud/umem/model/UMemDataSet.java
#	ucloud-sdk-java-umem/src/main/java/cn/ucloud/umem/model/UMemcachePriceSet.java
#	ucloud-sdk-java-umem/src/main/java/cn/ucloud/umem/model/URedisPriceSet.java
#	ucloud-sdk-java-unet/pom.xml
#	ucloud-sdk-java-unet/src/main/java/cn/ucloud/unet/model/CreateFirewallParam.java
#	ucloud-sdk-java-unet/src/main/java/cn/ucloud/unet/model/DescribeBandwidthUsageResult.java
#	ucloud-sdk-java-unet/src/main/java/cn/ucloud/unet/model/DescribeEIPResult.java
#	ucloud-sdk-java-unet/src/main/java/cn/ucloud/unet/model/DescribeFirewallResourceResult.java
#	ucloud-sdk-java-unet/src/main/java/cn/ucloud/unet/model/DescribeShareBandwidthResult.java
#	ucloud-sdk-java-unet/src/main/java/cn/ucloud/unet/model/GetEIPUpgradePriceResult.java
#	ucloud-sdk-java-usms/pom.xml
#	ucloud-sdk-java-usms/src/main/java/cn/ucloud/usms/client/DefaultUSMSClient.java
#	ucloud-sdk-java-usms/src/main/java/cn/ucloud/usms/client/USMSClient.java
#	ucloud-sdk-java-usms/src/main/java/cn/ucloud/usms/model/ReceiptPerPhone.java
#	ucloud-sdk-java-usms/src/main/java/cn/ucloud/usms/model/SendUSMSMessageParam.java
#	ucloud-sdk-java-usms/src/test/java/cn/ucloud/usms/client/SendUSMSMessageTest.java
#	ucloud-sdk-java-usql/pom.xml
#	ucloud-sdk-java-usql/src/main/java/cn/ucloud/usql/model/ListNamedQueriesParam.java
#	ucloud-sdk-java-usql/src/main/java/cn/ucloud/usql/model/ListSQLQueriesParam.java
#	ucloud-sdk-java-vpc/pom.xml
#	ucloud-sdk-java-vpc/src/main/java/cn/ucloud/vpc/model/DescribeRouteTableResult.java
#	ucloud-sdk-java-vpc/src/main/java/cn/ucloud/vpc/model/DescribeVPCResult.java

+++ ucloud-sdk-java-pathx/src/main/java/cn/ucloud/pathx/model/GetCommonUGAPriceResult.java
@@ -3,6 +3,8 @@ package cn.ucloud.pathx.model;
 import cn.ucloud.common.pojo.BaseResponseResult;
 import com.google.gson.annotations.SerializedName;
 
+import java.math.BigDecimal;
+
 
 /**
  * @Description : 获取苹果加速通道价格 结果类
@@ -14,14 +16,14 @@ public class GetCommonUGAPriceResult extends BaseResponseResult {
      * 价格
      */
     @SerializedName("Price")
-    private Integer price;
+    private BigDecimal price;
 
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return this.price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
